requirements.txt
streamlit
yfinance
python-dotenv
requests
groq
plotly
pandas
google-generativeai